follow pytest convention to actually execute some tests
